requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.5
django-cleanup==1.0.1
django-cors-headers==2.1.0
django-filter==1.0.4
django-rest-swagger==2.1.2
django-tenant-schemas==1.9.0
djangorestframework==3.6.4
djangorestframework-jwt==1.11.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
MonthDelta==0.9.1
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
psycopg2==2.7.3.1
py==1.4.34
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
simplejson==3.11.1
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
amqp==2.2.2
billiard==3.5.0.3
celery==4.1.0
django-celery-monitor==1.1.2
django-celery-results==1.0.1
django-tenants==1.3.3
kombu==4.1.0
redis==2.10.6
tenant-schemas-celery==0.1.7
vine==1.1.4
